requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
ediblepickle
hyperopt
matplotlib
seaborn
pandas
numpy
keras
scikit-learn
catboost
xlrd
imutils
#RUN ln -fs /usr/share/zoneinfo/Europe/Riga /etc/localtime
#RUN DEBIAN_FRONTEND=noninteractive apt-get install -y tzdata
pillow
#fastparquet
cmake
MulticoreTSNE
PyQt5
ipyparallel
opencv-python>=4.2.0.32
sacred
pymongo

#DARTS
numpy==1.15.4
graphviz==0.8.4
torch==1.5.1
torchvision==0.2.1
tensorboard==2.1.0
tensorboardX==1.6
